+++ src/runtime/src/ark/function.rs
@@ -1,4 +1,4 @@
-use crate::byte_utility::{split, split_as_u32};
+use crate::utility::bytes::{split, split_as_u32};
 use super::opcode::{Address, FuncId};
 
 /// Represents a function section.
