+++ src/runtime/src/ark/strings.rs
@@ -1,6 +1,6 @@
 use std::string::FromUtf8Error;
 
-use crate::byte_utility::{split, split_as_u32};
+use crate::utility::bytes::{split, split_as_u32};
 
 /// Represents a string section.
 #[derive(Debug, PartialEq, Eq)]
